More logical test grouping; more tests
- Grouped tests by modules
- Added tests for game mode and status effect modules
